# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BOTH & CO. ecommerce API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product management
  - name: cart
    description: Shopping cart
  - name: orders
    description: Order management
  - name: faq
    description: FAQ entries
  - name: services
    description: Services offered
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current user cart
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CartItem"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "201":
          description: Added cart item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear entire cart
      responses:
        "204":
          description: Cart cleared

  /cart/{id}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current user
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}/payment-proof:
    post:
      operationId: submitPaymentProof
      tags: [orders]
      summary: Submit payment proof for an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentProofInput"
      responses:
        "200":
          description: Proof submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /faq:
    get:
      operationId: listFaq
      tags: [faq]
      summary: List all FAQ entries
      responses:
        "200":
          description: FAQ list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FaqEntry"
    post:
      operationId: createFaqEntry
      tags: [faq]
      summary: Create FAQ entry (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqEntryInput"
      responses:
        "201":
          description: Created FAQ entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaqEntry"

  /faq/{id}:
    patch:
      operationId: updateFaqEntry
      tags: [faq]
      summary: Update FAQ entry (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqEntryUpdate"
      responses:
        "200":
          description: Updated FAQ entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaqEntry"
    delete:
      operationId: deleteFaqEntry
      tags: [faq]
      summary: Delete FAQ entry (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: Services list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get dashboard stats (admin)
      responses:
        "200":
          description: Admin dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List all orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /admin/orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /admin/products:
    get:
      operationId: listAdminProducts
      tags: [admin]
      summary: List all products for admin management
      responses:
        "200":
          description: Products list for admin
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      required: [id, name, price, category, inStock, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        category:
          type: string
          enum: [heels, glasses]
        imageUrl:
          type: ["string", "null"]
        inStock:
          type: boolean
        featured:
          type: boolean
        sizes:
          type: ["string", "null"]
        createdAt:
          type: string

    ProductInput:
      type: object
      required: [name, price, category]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
          enum: [heels, glasses]
        imageUrl:
          type: string
        inStock:
          type: boolean
        featured:
          type: boolean
        sizes:
          type: string

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
          enum: [heels, glasses]
        imageUrl:
          type: string
        inStock:
          type: boolean
        featured:
          type: boolean
        sizes:
          type: string

    CartItem:
      type: object
      required: [id, productId, quantity, sessionId]
      properties:
        id:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        sessionId:
          type: string
        size:
          type: ["string", "null"]
        product:
          $ref: "#/components/schemas/Product"

    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        size:
          type: string
        sessionId:
          type: string

    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer

    Order:
      type: object
      required: [id, status, totalAmount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["string", "null"]
        guestEmail:
          type: ["string", "null"]
        guestName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, payment_uploaded, confirmed, shipped, delivered, cancelled]
        totalAmount:
          type: number
        paymentProofUrl:
          type: ["string", "null"]
        shippingAddress:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    OrderItem:
      type: object
      required: [id, productId, quantity, price]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        price:
          type: number
        size:
          type: ["string", "null"]
        product:
          $ref: "#/components/schemas/Product"

    OrderInput:
      type: object
      required: [totalAmount, shippingAddress, phone]
      properties:
        guestEmail:
          type: string
        guestName:
          type: string
        totalAmount:
          type: number
        shippingAddress:
          type: string
        phone:
          type: string
        notes:
          type: string
        sessionId:
          type: string
        items:
          type: array
          items:
            type: object
            required: [productId, quantity, price]
            properties:
              productId:
                type: integer
              quantity:
                type: integer
              price:
                type: number
              size:
                type: string

    PaymentProofInput:
      type: object
      required: [paymentProofUrl]
      properties:
        paymentProofUrl:
          type: string

    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, payment_uploaded, confirmed, shipped, delivered, cancelled]

    FaqEntry:
      type: object
      required: [id, question, answer, order]
      properties:
        id:
          type: integer
        question:
          type: string
        answer:
          type: string
        order:
          type: integer

    FaqEntryInput:
      type: object
      required: [question, answer]
      properties:
        question:
          type: string
        answer:
          type: string
        order:
          type: integer

    FaqEntryUpdate:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
        order:
          type: integer

    Service:
      type: object
      required: [id, title, description]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        icon:
          type: ["string", "null"]

    AdminStats:
      type: object
      required: [totalOrders, pendingOrders, confirmedOrders, totalRevenue, totalProducts]
      properties:
        totalOrders:
          type: integer
        pendingOrders:
          type: integer
        confirmedOrders:
          type: integer
        totalRevenue:
          type: number
        totalProducts:
          type: integer
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/Order"