# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BOTH & CO. ecommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from './order';

export interface AdminStats {
  totalOrders: number;
  pendingOrders: number;
  confirmedOrders: number;
  totalRevenue: number;
  totalProducts: number;
  recentOrders?: Order[];
}