# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BOTH & CO. ecommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface CartItem {
  id: number;
  productId: number;
  quantity: number;
  sessionId: string;
  /** @nullable */
  size?: string | null;
  product?: Product;
}