# lib/api-zod/src/generated/types/faqEntryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BOTH & CO. ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface FaqEntryUpdate {
  question?: string;
  answer?: string;
  order?: number;
}