# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BOTH & CO. ecommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  guestEmail?: string | null;
  /** @nullable */
  guestName?: string | null;
  status: OrderStatus;
  totalAmount: number;
  /** @nullable */
  paymentProofUrl?: string | null;
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  items?: OrderItem[];
}