# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BOTH & CO. ecommerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputItemsItem } from './orderInputItemsItem';

export interface OrderInput {
  guestEmail?: string;
  guestName?: string;
  totalAmount: number;
  shippingAddress: string;
  phone: string;
  notes?: string;
  sessionId?: string;
  items?: OrderInputItemsItem[];
}